+++ testing/suites/config_discovery.json
@@ -6,38 +6,38 @@
       "name": "no_config_found",
       "description": "Error when no config file is found",
       "command": ["generate"],
-      "expected_err": "no airules configuration file found",
+      "expected_err": "no configuration file found",
       "exit_code": 1
     },
     {
       "name": "finds_airules_yaml",
-      "description": "Finds airules.yaml in current directory",
+      "description": "Finds ai-rulez.yaml in current directory",
       "command": ["generate", "--dry-run"],
       "files": {
-        "airules.yaml": "metadata:\n  name: \"Test Project\"\noutputs:\n  - file: \"test.md\"\nrules:\n  - name: \"Test Rule\"\n    content: \"Test content\""
+        "ai-rulez.yaml": "metadata:\n  name: \"Test Project\"\noutputs:\n  - file: \"test.md\"\nrules:\n  - name: \"Test Rule\"\n    content: \"Test content\""
       },
-      "expected_out": "Using config: airules.yaml",
+      "expected_out": "Using config file:",
       "exit_code": 0
     },
     {
       "name": "finds_dot_airules_yaml",
-      "description": "Finds .airules.yaml in current directory",
+      "description": "Finds .ai-rulez.yaml in current directory",
       "command": ["generate", "--dry-run"],
       "files": {
-        ".airules.yaml": "metadata:\n  name: \"Test Project\"\noutputs:\n  - file: \"test.md\"\nrules:\n  - name: \"Test Rule\"\n    content: \"Test content\""
+        ".ai-rulez.yaml": "metadata:\n  name: \"Test Project\"\noutputs:\n  - file: \"test.md\"\nrules:\n  - name: \"Test Rule\"\n    content: \"Test content\""
       },
-      "expected_out": "Using config: .airules.yaml",
+      "expected_out": "Using config file:",
       "exit_code": 0
     },
     {
       "name": "prefers_dot_airules_yaml",
-      "description": "Prefers .airules.yaml over airules.yaml",
+      "description": "Prefers .ai-rulez.yaml over ai-rulez.yaml",
       "command": ["generate", "--dry-run"],
       "files": {
-        ".airules.yaml": "metadata:\n  name: \"Dot Config\"\noutputs:\n  - file: \"test.md\"\nrules:\n  - name: \"Test Rule\"\n    content: \"Test content\"",
-        "airules.yaml": "metadata:\n  name: \"Plain Config\"\noutputs:\n  - file: \"test.md\"\nrules:\n  - name: \"Test Rule\"\n    content: \"Test content\""
+        ".ai-rulez.yaml": "metadata:\n  name: \"Dot Config\"\noutputs:\n  - file: \"test.md\"\nrules:\n  - name: \"Test Rule\"\n    content: \"Test content\"",
+        "ai-rulez.yaml": "metadata:\n  name: \"Plain Config\"\noutputs:\n  - file: \"test.md\"\nrules:\n  - name: \"Test Rule\"\n    content: \"Test content\""
       },
-      "expected_out": "Using config: .airules.yaml",
+      "expected_out": "Using config file:",
       "exit_code": 0
     },
     {
@@ -47,9 +47,9 @@
       "command": ["generate", "--dry-run"],
       "setup": ["mkdir -p subdir"],
       "files": {
-        "airules.yaml": "metadata:\n  name: \"Parent Config\"\noutputs:\n  - file: \"test.md\"\nrules:\n  - name: \"Test Rule\"\n    content: \"Test content\""
+        "ai-rulez.yaml": "metadata:\n  name: \"Parent Config\"\noutputs:\n  - file: \"test.md\"\nrules:\n  - name: \"Test Rule\"\n    content: \"Test content\""
       },
-      "expected_out": "Using config: ../airules.yaml",
+      "expected_out": "Using config file:",
       "exit_code": 0
     },
     {
@@ -58,8 +58,8 @@
       "command": ["generate", "--recursive", "--dry-run"],
       "setup": ["mkdir -p project1", "mkdir -p project2"],
       "files": {
-        "project1/.airules.yaml": "metadata:\n  name: \"Project 1\"\noutputs:\n  - file: \"test1.md\"\nrules:\n  - name: \"Rule 1\"\n    content: \"Content 1\"",
-        "project2/airules.yaml": "metadata:\n  name: \"Project 2\"\noutputs:\n  - file: \"test2.md\"\nrules:\n  - name: \"Rule 2\"\n    content: \"Content 2\""
+        "project1/.ai-rulez.yaml": "metadata:\n  name: \"Project 1\"\noutputs:\n  - file: \"test1.md\"\nrules:\n  - name: \"Rule 1\"\n    content: \"Content 1\"",
+        "project2/ai-rulez.yaml": "metadata:\n  name: \"Project 2\"\noutputs:\n  - file: \"test2.md\"\nrules:\n  - name: \"Rule 2\"\n    content: \"Content 2\""
       },
       "expected_out": "Found 2 configuration file(s) for validation",
       "exit_code": 0
@@ -71,7 +71,7 @@
       "files": {
         "custom-config.yaml": "metadata:\n  name: \"Custom Config\"\noutputs:\n  - file: \"custom.md\"\nrules:\n  - name: \"Custom Rule\"\n    content: \"Custom content\""
       },
-      "expected_out": "Configuration valid",
+      "expected_out": "DRY RUN MODE",
       "exit_code": 0
     },
     {
