+++ testing/suites/generation.json
@@ -6,35 +6,35 @@
       "name": "basic_generation",
       "description": "Generate files from basic configuration",
       "command": ["generate", "scenarios/basic/ai_rulez.yaml"],
-      "expected_out": "Generated * output file(s)",
+      "expected_out": "Generated * file(s) successfully",
       "exit_code": 0
     },
     {
       "name": "dry_run_mode",
       "description": "Dry run shows what would be generated",
       "command": ["generate", "scenarios/basic/ai_rulez.yaml", "--dry-run"],
-      "expected_out": "Configuration valid. Would generate * file(s)",
+      "expected_out": "Would generate * output file(s)",
       "exit_code": 0
     },
     {
       "name": "minimal_config",
       "description": "Generate from minimal configuration",
       "command": ["generate", "scenarios/minimal/ai_rulez.yaml"],
-      "expected_out": "Generated * output file(s)",
+      "expected_out": "Generated * file(s) successfully",
       "exit_code": 0
     },
     {
       "name": "with_includes",
       "description": "Generate with included files",
       "command": ["generate", "scenarios/with-includes/ai_rulez.yaml"],
-      "expected_out": "Generated * output file(s)",
+      "expected_out": "Generated * file(s) successfully",
       "exit_code": 0
     },
     {
       "name": "nested_includes",
       "description": "Generate with nested includes",
       "command": ["generate", "scenarios/nested-includes/ai_rulez.yaml"],
-      "expected_out": "Generated * output file(s)",
+      "expected_out": "Generated * file(s) successfully",
       "exit_code": 0
     },
     {
@@ -47,7 +47,7 @@
       "setup": [
         "mkdir -p outputs"
       ],
-      "expected_out": "Generated * output file(s)",
+      "expected_out": "Generated * file(s) successfully",
       "exit_code": 0
     },
     {
@@ -58,7 +58,7 @@
         "custom-template.yaml": "metadata:\n  name: \"Custom Template Test\"\noutputs:\n  - file: \"custom.md\"\n    template: \"@templates/simple.tmpl\"\nrules:\n  - name: \"Test Rule\"\n    content: \"Test content\"",
         "templates/simple.tmpl": "{{.ProjectName}} Rules\n\n{{range .Rules}}\n{{.Name}}: {{.Content}}\n{{end}}"
       },
-      "expected_out": "Generated * output file(s)",
+      "expected_out": "Generated * file(s) successfully",
       "exit_code": 0
     },
     {
@@ -68,7 +68,7 @@
       "files": {
         "inline-template.yaml": "metadata:\n  name: \"Inline Template Test\"\noutputs:\n  - file: \"inline.md\"\n    template: |\n      # {{.ProjectName}} Rules\n      {{range .Rules}}\n      ## {{.Name}}\n      {{.Content}}\n      {{end}}\nrules:\n  - name: \"Test Rule\"\n    content: \"Test content\""
       },
-      "expected_out": "Generated * output file(s)",
+      "expected_out": "Generated * file(s) successfully",
       "exit_code": 0
     },
     {
@@ -78,7 +78,7 @@
       "files": {
         "relative-test.yaml": "metadata:\n  name: \"Relative Path Test\"\noutputs:\n  - file: \"docs/rules.md\"\n  - file: \"config/.cursorrules\"\nrules:\n  - name: \"Test Rule\"\n    content: \"Test content\""
       },
-      "expected_out": "Generated * output file(s)",
+      "expected_out": "Generated * file(s) successfully",
       "exit_code": 0
     }
   ]
